Generate provider declarations
